refactor program-lib
